refactor: restructure components into 'section' directory
- Moved components from `src/components` to `src/section` directory.
- Updated all import paths to reflect the new directory structure.
- Adjusted test files to use the updated component paths.
